@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.black]
 line_length = 79
-target-version = ["py36"]
+target-version = ["py37"]
 
 [tool.isort]
 profile = "black"
@@ -18,6 +18,7 @@ known_first_party = [
 ]
 combine_as_imports = true
 lines_between_types = 1
+py_version=37
 
 [tool.coverage.run]
 data_file = "reports/.coverage"
