@@ -17,7 +17,6 @@ project_urls =
 classifiers =
     Intended Audience :: Developers
     Topic :: Software Development :: Libraries
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -34,7 +33,7 @@ keywords =
     dailymotion
 
 [options]
-python_requires = >=3.6
+python_requires = >=3.7
 install_requires =
     cffi>=1.0.0,<2.0.0
     lark>=1.0.0,<2.0.0
